+++ test/wrappers/ortools_test.rb
@@ -3222,6 +3222,27 @@ class Wrappers::OrtoolsTest < Minitest::Test
     assert(solution.routes.any?{ |route| route.stops.last.service_id == 'service_3' })
   end
 
+  def test_exclusive
+    problem = VRP.basic
+    problem[:vehicles] << problem[:vehicles][0].dup
+    problem[:vehicles][1][:id] += '_copy'
+    problem[:vehicles].each{ |v| v[:start_point_id] = nil }
+    problem[:relations] = [{
+      type: :exclusive,
+      linked_ids: ['service_1', 'service_3']
+    }]
+
+    vrp = TestHelper.create(problem)
+    assert_equal :exclusive, vrp.services[0].activity.position
+    assert_equal :neutral, vrp.services[1].activity.position
+    assert_equal :exclusive, vrp.services[2].activity.position
+    solution = OptimizerWrapper.config[:services][:ortools].solve(vrp, 'test')
+    routes_with_s1 = solution.routes.find{ |r| r.stops.any?{ |s| s.service_id == "service_1" } }
+    routes_with_s3 = solution.routes.find{ |r| r.stops.any?{ |s| s.service_id == "service_3" } }
+    assert_equal routes_with_s1.vehicle.id, routes_with_s3.vehicle.id
+    assert_equal 2, routes_with_s1.stops.size
+  end
+
   def test_never_first
     problem = VRP.basic
     problem[:vehicles] << problem[:vehicles][0].dup